Update offers; Known issue: When updating validation is not working

+++ SpringFundamentals/SF02_SpringApp-MobiLeLeLe/src/main/java/bg/softuni/mobilelele/model/entity/OfferEntity.java
@@ -30,7 +30,7 @@ public class OfferEntity extends BaseEntity {
     private String imageUrl;
 
     @Column
-    private int mileage;
+    private double mileage;
 
     @Column
     private BigDecimal price;
